+++ library/src/main/kotlin/de/adesso/junitinsights/autoconfiguration/InsightReportProperties.kt
@@ -3,7 +3,7 @@ package de.adesso.junitinsights.autoconfiguration
 import org.springframework.boot.context.properties.ConfigurationProperties
 
 @ConfigurationProperties(prefix = "junitinsights.report")
-object JUnitInsightsReportProperties {
+object InsightReportProperties {
     var path: String = ""
     val templatepath: String = "/htmlTemplate.html"
 }
