+++ library/src/main/resources/META-INF/services/org.junit.jupiter.api.extension.Extension
@@ -1,1 +1,1 @@
-de.adesso.junitinsights.tests.TestBenchmarkExtension
+de.adesso.junitinsights.junit.InsightExtension
