+++ library/src/main/resources/META-INF/services/org.junit.platform.launcher.TestExecutionListener
@@ -1,1 +1,1 @@
-de.adesso.junitinsights.tests.InsightListener
+de.adesso.junitinsights.junit.InsightListener
